[nrf fromlist] drivers: serial: nrfx_uarte: Add runtime PM
Add runtime PM to the driver. When asynchronous or interrupt
driven API is used, there are 3 independent paths for enabling
the device: RX, TX and TX poll_out. TX poll_out requires special
handling because number of poll_out calls does not need to much
number of TXSTOPPED interrupts. To handle that special flag is
added. For standard RX and TX is simpler.

https://github.com/zephyrproject-rtos/zephyr/pull/80201

Signed-off-by: Krzysztof Chruściński <krzysztof.chruscinski@nordicsemi.no>