[nrf fromlist] drivers: serial: nrfx_uarte: Improve RX FIFO workaround
Current workaround fails to detect data in FIFO if the last received
data was equal to the magic byte used in the workaround and length
of the transfer was short (less than HW FIFO size which is 5). In order
to reduce chances of failure additional measure is added. Data is
added to the FIFO if some byte arrives after ENDRX. Each byte triggers
RXDRDY event. We attempt to clear RXDRDY event as one of the first
things when entering ISR and then when RX FIFO is flushed RXDRDY event
is checked. If it is set, it means that there was a byte coming after
ENDRX so FIFO was not empty. Workaround will fail in case when ISR
latency is long enough that byte to FIFO comes before ISR is handled.

Upstream PR: https://github.com/zephyrproject-rtos/zephyr/pull/80201

Signed-off-by: Krzysztof Chruściński <krzysztof.chruscinski@nordicsemi.no>