[nrf fromlist] drivers: serial: nrfx_uarte: Perform flushing only when disabling
Previously, RX FIFO flushing was performed after each RXTO
(disabling of RX). However, it is necessary only when whole
UARTE is disabled. It is disabled when there is no RX or TX
active or when runtime device PM is enabled it is disabled
when there is no one requesting UARTE to be active. In case
of PM it is possible that UARTE will be kept powered up for
longer period of time and in that case flushing is not
performed. Because currently UARTE in all Nordic SoCs has
a bug in the RX FIFO we would like to avoid flushing if
possible since workaround is not covering 100% of cases.

Because UARTE is disabling is performed with interrupts
locked we try to limit number of operations there. RX FIFO
flushing is split into preparation (rx_flush_prepare())
which is performed in every RXTO event handling and actual
flush (rx_flush()) which is performed only when UARTE is
disabled.

Upstream PR: https://github.com/zephyrproject-rtos/zephyr/pull/80201

Signed-off-by: Krzysztof Chruściński <krzysztof.chruscinski@nordicsemi.no>